requirements.txt
google-generativeai
python-dotenv
streamlit
youtube-transcript-api
pathlib